refactor: Break out romeo actors to separate modules
