--- romeo/ContractManager.json
@@ -1,4 +1,0 @@
-{
-  "deployed": true,
-  "contract": null
-}
