--- romeo/DepositProcessor.json
@@ -1,5 +1,0 @@
-{
-  "block_height": 0,
-  "next_nonce": 0,
-  "deposits": {}
-}
